完成锤子

+++ src/main/kotlin/cn/solarmoon/spirit_of_fight/feature/fight_skill/spirit/FightSpirit.kt
@@ -35,7 +35,7 @@ class FightSpirit {
         }
     }
 
-    fun addStage(multiplier: Float) = addStage((baseGrowth * multiplier).toInt())
+    fun addStage(multiplier: Double) = addStage((baseGrowth * multiplier).toInt())
 
     fun clear() {
         valueCache = value
